backend/requirements.txt
fastapi
uvicorn
authlib
httpx
python-dotenv
cryptography
starlette
itsdangerous
python-multipart
email-validator
openai
supabase
